Impl. user config and asst. command fixes
* $XDG_CONFIG_HOME/bak.cfg now exists
* bakfile selection now supports diff, view
* Prompts now handle uppercase
* Most failures are now graceful

+++ bak/commands/__init__.py
@@ -9,6 +9,7 @@ from typing import List
 from warnings import warn
 
 import click
+from config import Config
 
 from bak.data import bakfile, bak_db
 
@@ -22,8 +23,13 @@ try:
     config_dir = os.environ["XDG_CONFIG_HOME"]
 except KeyError:
     config_dir = os.path.expanduser("~/.config")
-bak_dir = os.path.join(data_dir, "bak", "bakfiles")
-bak_db_loc = os.path.join(data_dir, "bak", "bak.db")
+
+config_file = os.path.join(config_dir, 'bak.cfg')
+cfg = Config(config_file)
+
+bak_dir = cfg['bakfile_location'] or os.path.join(data_dir, "bak", "bakfiles")
+bak_db_loc = cfg['bak_database_location'] or \
+    os.path.join(data_dir, "bak", "bak.db")
 
 if not os.path.exists(bak_dir):
     os.makedirs(bak_dir)
@@ -40,12 +46,11 @@ def _assemble_bakfile(filename):
     splitname = os.path.split(expandpath(filename))
     bakfile_name = "".join([".".join(i[1:].replace("/", "-") for i in splitname[:-1]) +
                             '-' +
-                            splitname[-1],  # [os.path.split(filename)[-1],
+                            splitname[-1],
                             ".",
                             '-'.join(str(
                                 time_now.timestamp()).split('.')),
                             ".bak"]).replace(" ", "-")
-    # TODO #26 get bakfile directory from config
     bakfile_path = os.path.join(bak_dir, bakfile_name)
 
     new_bak_entry = bakfile.BakFile(os.path.basename(filename),
@@ -56,7 +61,7 @@ def _assemble_bakfile(filename):
     return new_bak_entry
 
 
-default_select_prompt = ("Enter a number, or: (V)iew (C)ancel", 'c')
+default_select_prompt = ("Enter a number, or: (V)iew (D)iff (C)ancel", 'c')
 
 
 def _get_bakfile_entry(filename, select_prompt=default_select_prompt, err=False):
@@ -76,33 +81,44 @@ def _do_select_bakfile(bakfiles: List[bakfile.BakFile],
     for i in _range:
         click.echo(
             f"{i + 1}: .bakfile last modified at {bakfiles[i].date_modified}", err=err)
-    choice = click.prompt(*select_prompt, err=err)
-    # TODO add diff and print as choices here
-    # "Enter a number, or: (V)iew (C)ancel", default='c').lower()
-    if choice != "c":
-        view = False
-        try:
-            if choice == "v":
-                idx = int(click.prompt("View which .bakfile?", err=err)) - 1
-                # idx = int(click.prompt("View which .bakfile? (#)")) - 1
-                view = True
-            else:
-                idx = int(choice) - 1
-            if idx not in _range:
-                click.echo("Invalid selection. Aborting.", err=err)
-                return None
-            else:
-                if view:
-                    bak_print_cmd(bakfiles[idx])
-                    return
-                return bakfiles[idx]
-        except (ValueError, TypeError) as e:
-            warn(e)
-            click.echo("Invalid input. Aborting.", err=err)
+
+    def get_choice():
+        return click.prompt(*select_prompt, err=err).lower()
+    choice = get_choice()
+
+    while True:
+        if choice == "c":
+            click.echo("Cancelled.", err=err)
             return None
-    else:
-        click.echo("Aborting.", err=err)
-        return None
+        else:
+            view = False
+            try:
+                if choice == "v":
+                    idx = int(click.prompt(
+                        "View which .bakfile?", err=err)) - 1
+                    view = True
+                elif choice == "d":
+                    idx = int(click.prompt(
+                        "Diff which .bakfile?", err=err)) - 1
+                    bak_diff_cmd(bakfiles[idx])
+                    choice = get_choice()
+                    continue
+                else:
+                    idx = int(choice) - 1
+                if idx not in _range:
+                    click.echo("Invalid selection. Aborting.", err=err)
+                    return None
+                elif view:
+                    bak_print_cmd(bakfiles[idx])
+                    choice = get_choice()
+                    continue
+                else:
+                    return bakfiles[idx]
+            except (ValueError, TypeError) as e:
+                warn(e)
+                click.echo("Invalid input. Aborting.", err=err)
+                return None
+            get_choice()
 
 
 def show_bak_list(filename: (None, str, os.path) = None):
@@ -140,7 +156,9 @@ def bak_up_cmd(filename: str):
         filename (str|os.path)
     """
     # Return Truthy things for failures that echo their own output,
-    # false for nonspecific or generic failures
+    # false for nonspecific or generic failures.
+    # Put differently, False is for complete failures. If this function
+    # handles a failure gracefully, it should return True.
 
     filename = expandpath(filename)
     old_bakfile = db_handler.get_bakfile_entries(filename)
@@ -163,7 +181,8 @@ def bak_up_cmd(filename: str):
 
 
 def bak_down_cmd(filename: str,
-                 keep_bakfile: bool = False):
+                 keep_bakfile: bool = False,
+                 quiet: bool = False):
     """ Restore `filename` from .bakfile. Prompts if ambiguous (such as
         when there are multiple .bakfiles of `filename`)
 
@@ -173,14 +192,25 @@ def bak_down_cmd(filename: str,
     """
     filename = expandpath(filename)
     bakfile_entries = db_handler.get_bakfile_entries(filename)
+    if not bakfile_entries:
+        click.echo(f"No bakfiles found for {filename}")
+        return
 
-    # TODO still only pulling first result
     bakfile_entry = _do_select_bakfile(bakfile_entries) if len(
         bakfile_entries) > 1 else bakfile_entries[0]
 
     if not bakfile_entry:
+        click.echo(f"No bakfiles found for {filename}")
         return
 
+    confirm_prompt = f"Confirm: Restore {filename} and erase bakfiles?\n" \
+                     if not keep_bakfile else \
+                     f"Confirm: Restore {filename} and keep bakfiles?\n"
+    confirm_prompt += "(y/n)"
+    confirm = click.prompt(confirm_prompt, default='y')
+    if confirm.lower()[0] != 'y':
+        click.echo("Cancelled.")
+        return
     os.remove(filename)
     copy2(bakfile_entry.bakfile_loc, filename)
     if not keep_bakfile:
@@ -212,7 +242,8 @@ def bak_off_cmd(filename: (None, str, os.path),
     filename = expandpath(filename)
     bakfiles = db_handler.get_bakfile_entries(filename)
     if not bakfiles:
-        click.echo(f"No bakfiles found for {filename}")
+        click.echo(f"No bakfiles found for {os.path.abspath(filename)}")
+        return False
     confirm = input(
         f"Confirming: Remove {len(bakfiles)} .bakfiles for {filename}? "
         f"(y/N) ").lower() == 'y' if not quietly else True
@@ -227,28 +258,38 @@ def bak_print_cmd(bak_to_print: (str, bakfile.BakFile), using: (str, None) = Non
     if not isinstance(bak_to_print, bakfile.BakFile):
         bak_to_print = _get_bakfile_entry(bak_to_print,
                                           select_prompt=("View which .bakfile? (#)", "c"))
+        if not bak_to_print:
+            click.echo(f"No bakfiles found for {os.path.abspath(filename)}")
         if not isinstance(bak_to_print, bakfile.BakFile):
             return  # _get_bakfile_entry() handles failures, so just exit here
     if using:
         pager = using
     else:
-        try:
-            pager = os.environ['PAGER']
-        except KeyError:
-            pager = 'less'
+        # try:
+        pager = (cfg['bak_show_exec'] or os.environ['PAGER']) or 'less'
+        # except KeyError:
+        #     pager = 'less'
     call([pager, bak_to_print.bakfile_loc])
 
 
 def bak_getfile_cmd(bak_to_get: (str, bakfile.BakFile)):
     if not isinstance(bak_to_get, bakfile.BakFile):
+        filename = bak_to_get
         bak_to_get = _get_bakfile_entry(bak_to_get, err=True)
         if not bak_to_get:
+            click.echo(f"No bakfiles found for {os.path.abspath(filename)}")
             return  # _get_bakfile_entry() handles failures, so just exit
-        click.echo(bak_to_get.bakfile_loc)
+    click.echo(bak_to_get.bakfile_loc)
 
 
-def bak_diff_cmd(filename, command='diff'):
+def bak_diff_cmd(filename: str, command='diff'):
     # TODO write tests for this (mildly tricky)
-    bak_to_diff = _get_bakfile_entry(expandpath(filename))
-    call([command if command else 'diff',
+    bak_to_diff = filename if isinstance(
+        filename, bakfile.BakFile) else _get_bakfile_entry(expandpath(filename))
+    if not command:
+        command = cfg['bak_diff_exec'] or 'diff'
+    if not bak_to_diff:
+        click.echo(f"No bakfiles found for {os.path.abspath(filename)}")
+        return
+    call([command,
           bak_to_diff.bakfile_loc, bak_to_diff.orig_abspath])
